fix(WebGPU): fix a texture issue on WebGPU
Another issue in the texture copy/resize code. Seems to
work now.

Also update some tests locations and make them run on
webgpu as well.
