+++ Sources/Rendering/Core/ColorTransferFunction/test/testColorTransferFunctionPresets.js
@@ -13,9 +13,9 @@ import baseline from './testColorTransferFunctionPresets.png';
 const MAX_NUMBER_OF_PRESETS = 200;
 const NUMBER_PER_LINE = 20;
 
-test('Test Interpolate Scalars Before Colors', (t) => {
+test('Test ColorTransferFunction Presets', (t) => {
   const gc = testUtils.createGarbageCollector(t);
-  t.ok('rendering', 'vtkOpenGLPolyDataMapper ColorTransferFunction Presets');
+  t.ok('rendering', 'vtkMapper ColorTransferFunction Presets');
 
   // Create some control UI
   const container = document.querySelector('body');
