+++ Sources/Rendering/Core/VolumeMapper/test/testAverageIntensityProjection.js
@@ -3,20 +3,20 @@ import testUtils from 'vtk.js/Sources/Testing/testUtils';
 
 import vtkColorTransferFunction from 'vtk.js/Sources/Rendering/Core/ColorTransferFunction';
 import vtkHttpDataSetReader from 'vtk.js/Sources/IO/Core/HttpDataSetReader';
-import vtkOpenGLRenderWindow from 'vtk.js/Sources/Rendering/OpenGL/RenderWindow';
+import 'vtk.js/Sources/Rendering/Misc/RenderingAPIs';
 import vtkPiecewiseFunction from 'vtk.js/Sources/Common/DataModel/PiecewiseFunction';
 import vtkRenderWindow from 'vtk.js/Sources/Rendering/Core/RenderWindow';
 import vtkRenderWindowInteractor from 'vtk.js/Sources/Rendering/Core/RenderWindowInteractor';
 import vtkRenderer from 'vtk.js/Sources/Rendering/Core/Renderer';
 import vtkVolume from 'vtk.js/Sources/Rendering/Core/Volume';
 import vtkVolumeMapper from 'vtk.js/Sources/Rendering/Core/VolumeMapper';
+import Constants from 'vtk.js/Sources/Rendering/Core/VolumeMapper/Constants';
 
-import baseline1 from './testComposite16Bit.png';
-import baseline2 from './testComposite16Bit_2.png';
+import baseline from './testAverageIntensityProjection.png';
 
-test.onlyIfWebGL('Test Composite Volume Rendering', (t) => {
+test('Test Average Intensity Projection Volume Rendering', (t) => {
   const gc = testUtils.createGarbageCollector(t);
-  t.ok('rendering', 'vtkOpenGLVolumeMapper Composite16Bit');
+  t.ok('rendering', 'vtkVolumeMapper AverageIP');
   // testUtils.keepDOM();
 
   // Create some control UI
@@ -30,35 +30,41 @@ test.onlyIfWebGL('Test Composite Volume Rendering', (t) => {
   const renderWindow = gc.registerResource(vtkRenderWindow.newInstance());
   const renderer = gc.registerResource(vtkRenderer.newInstance());
   renderWindow.addRenderer(renderer);
+  renderer.setBackground(0.32, 0.34, 0.43);
 
   const actor = gc.registerResource(vtkVolume.newInstance());
-  // renderer.addVolume(actor);
 
   const mapper = gc.registerResource(vtkVolumeMapper.newInstance());
-  mapper.setSampleDistance(2.0);
+  mapper.setSampleDistance(0.7);
+  mapper.setBlendMode(Constants.BlendMode.AVERAGE_INTENSITY_BLEND);
+
   actor.setMapper(mapper);
+
+  const reader = vtkHttpDataSetReader.newInstance({ fetchGzip: true });
 
   // create color and opacity transfer functions
   const ctfun = vtkColorTransferFunction.newInstance();
-  ctfun.addRGBPoint(200.0, 1.0, 1.0, 1.0);
-  ctfun.addRGBPoint(2000.0, 0.0, 0.0, 0.0);
+  ctfun.addRGBPoint(-3024, 0, 0, 0);
+  ctfun.addRGBPoint(-637.62, 1, 1, 1);
+  ctfun.addRGBPoint(700, 1, 1, 1);
+  ctfun.addRGBPoint(3071, 1, 1, 1);
+  ctfun.setMappingRange(500, 3000);
+
   const ofun = vtkPiecewiseFunction.newInstance();
-  ofun.addPoint(200.0, 0.0);
-  ofun.addPoint(1200.0, 0.2);
-  ofun.addPoint(4000.0, 0.4);
+  ofun.addPoint(-3024, 0);
+  ofun.addPoint(-637.62, 0);
+  ofun.addPoint(700, 0.5);
+  ofun.addPoint(3071, 0.9);
+
   actor.getProperty().setRGBTransferFunction(0, ctfun);
   actor.getProperty().setScalarOpacity(0, ofun);
   actor.getProperty().setScalarOpacityUnitDistance(0, 4.5);
-  // actor.getProperty().setInterpolationTypeToNearest();
-  // actor.getProperty().setInterpolationTypeToFastLinear();
-  actor.getProperty().setInterpolationTypeToLinear();
+  actor.getProperty().setInterpolationTypeToFastLinear();
 
-  const reader = vtkHttpDataSetReader.newInstance({ fetchGzip: true });
-  mapper.setImageSampleDistance(2.0);
   mapper.setInputConnection(reader.getOutputPort());
 
-  // now create something to view it, in this case webgl
-  const glwindow = gc.registerResource(vtkOpenGLRenderWindow.newInstance());
+  // now create something to view it
+  const glwindow = gc.registerResource(renderWindow.newAPISpecificView());
   glwindow.setContainer(renderWindowContainer);
   renderWindow.addView(glwindow);
   glwindow.setSize(400, 400);
@@ -74,11 +80,12 @@ test.onlyIfWebGL('Test Composite Volume Rendering', (t) => {
     reader.loadData().then(() => {
       renderer.addVolume(actor);
       renderer.resetCamera();
+
       glwindow.captureNextImage().then((image) => {
         testUtils.compareImages(
           image,
-          [baseline1, baseline2],
-          'Rendering/OpenGL/VolumeMapper/testComposite16Bit',
+          [baseline],
+          'Rendering/Core/VolumeMapper/testAverageIntensityProjection',
           t,
           1.5,
           gc.releaseResources
