+++ Sources/Rendering/Core/VolumeMapper/test/testAverageIntensityProjection.png
