+++ Sources/Rendering/Core/VolumeMapper/test/testComposite.js
@@ -1,22 +1,22 @@
 import test from 'tape-catch';
 import testUtils from 'vtk.js/Sources/Testing/testUtils';
 
+import 'vtk.js/Sources/Rendering/Misc/RenderingAPIs';
 import vtkColorTransferFunction from 'vtk.js/Sources/Rendering/Core/ColorTransferFunction';
 import vtkHttpDataSetReader from 'vtk.js/Sources/IO/Core/HttpDataSetReader';
-import vtkOpenGLRenderWindow from 'vtk.js/Sources/Rendering/OpenGL/RenderWindow';
 import vtkPiecewiseFunction from 'vtk.js/Sources/Common/DataModel/PiecewiseFunction';
 import vtkRenderWindow from 'vtk.js/Sources/Rendering/Core/RenderWindow';
 import vtkRenderWindowInteractor from 'vtk.js/Sources/Rendering/Core/RenderWindowInteractor';
 import vtkRenderer from 'vtk.js/Sources/Rendering/Core/Renderer';
 import vtkVolume from 'vtk.js/Sources/Rendering/Core/Volume';
 import vtkVolumeMapper from 'vtk.js/Sources/Rendering/Core/VolumeMapper';
-import Constants from 'vtk.js/Sources/Rendering/Core/VolumeMapper/Constants';
 
-import baseline from './testMinimumIntensityProjection.png';
+import baseline1 from './testComposite.png';
+import baseline2 from './testComposite_2.png';
 
-test.onlyIfWebGL('Test Minimum Intensity Projection Volume Rendering', (t) => {
+test('Test Composite Volume Rendering', (t) => {
   const gc = testUtils.createGarbageCollector(t);
-  t.ok('rendering', 'vtkOpenGLVolumeMapper MinIP');
+  t.ok('rendering', 'vtkVolumeMapper Composite');
   // testUtils.keepDOM();
 
   // Create some control UI
@@ -30,31 +30,24 @@ test.onlyIfWebGL('Test Minimum Intensity Projection Volume Rendering', (t) => {
   const renderWindow = gc.registerResource(vtkRenderWindow.newInstance());
   const renderer = gc.registerResource(vtkRenderer.newInstance());
   renderWindow.addRenderer(renderer);
-  renderer.setBackground(0.3, 0.3, 0.3);
+  renderer.setBackground(0.32, 0.34, 0.43);
 
   const actor = gc.registerResource(vtkVolume.newInstance());
 
   const mapper = gc.registerResource(vtkVolumeMapper.newInstance());
-  mapper.setSampleDistance(1.3);
-  mapper.setBlendMode(Constants.BlendMode.MINIMUM_INTENSITY_BLEND);
-
+  mapper.setSampleDistance(0.7);
   actor.setMapper(mapper);
 
   const reader = vtkHttpDataSetReader.newInstance({ fetchGzip: true });
-
   // create color and opacity transfer functions
   const ctfun = vtkColorTransferFunction.newInstance();
-  ctfun.addRGBPoint(-3024, 1, 1, 1);
-  ctfun.addRGBPoint(-637.62, 1, 1, 1);
-  ctfun.addRGBPoint(700, 0, 0, 0);
-  ctfun.addRGBPoint(3071, 0, 0, 0);
-
+  ctfun.addRGBPoint(0, 85 / 255.0, 0, 0);
+  ctfun.addRGBPoint(95, 1.0, 1.0, 1.0);
+  ctfun.addRGBPoint(225, 0.66, 0.66, 0.5);
+  ctfun.addRGBPoint(255, 0.3, 1.0, 0.5);
   const ofun = vtkPiecewiseFunction.newInstance();
-  ofun.addPoint(-3024, 0.1);
-  ofun.addPoint(-637.62, 0.1);
-  ofun.addPoint(700, 0.9);
-  ofun.addPoint(3071, 0.9);
-
+  ofun.addPoint(0.0, 0.0);
+  ofun.addPoint(255.0, 1.0);
   actor.getProperty().setRGBTransferFunction(0, ctfun);
   actor.getProperty().setScalarOpacity(0, ofun);
   actor.getProperty().setScalarOpacityUnitDistance(0, 3.0);
@@ -62,8 +55,8 @@ test.onlyIfWebGL('Test Minimum Intensity Projection Volume Rendering', (t) => {
 
   mapper.setInputConnection(reader.getOutputPort());
 
-  // now create something to view it, in this case webgl
-  const glwindow = gc.registerResource(vtkOpenGLRenderWindow.newInstance());
+  // now create something to view it
+  const glwindow = gc.registerResource(renderWindow.newAPISpecificView());
   glwindow.setContainer(renderWindowContainer);
   renderWindow.addView(glwindow);
   glwindow.setSize(400, 400);
@@ -75,17 +68,19 @@ test.onlyIfWebGL('Test Minimum Intensity Projection Volume Rendering', (t) => {
   interactor.initialize();
   interactor.bindEvents(renderWindowContainer);
 
-  reader.setUrl(`${__BASE_PATH__}/Data/volume/headsq.vti`).then(() => {
+  reader.setUrl(`${__BASE_PATH__}/Data/volume/LIDC2.vti`).then(() => {
     reader.loadData().then(() => {
       renderer.addVolume(actor);
       renderer.resetCamera();
-      renderer.getActiveCamera().elevation(-120);
+      renderer.getActiveCamera().zoom(1.5);
+      renderer.getActiveCamera().elevation(70);
+      renderer.resetCameraClippingRange();
 
       glwindow.captureNextImage().then((image) => {
         testUtils.compareImages(
           image,
-          [baseline],
-          'Rendering/OpenGL/VolumeMapper/testMinimumIntensityProjection',
+          [baseline1, baseline2],
+          'Rendering/Core/VolumeMapper/testComposite',
           t,
           1.5,
           gc.releaseResources
