+++ Sources/Rendering/Core/VolumeMapper/test/testComposite.png
