+++ Sources/Rendering/Core/VolumeMapper/test/testComposite16Bit.png
