+++ Sources/Rendering/Core/VolumeMapper/test/testComposite16Bit_2.png
