+++ Sources/Rendering/Core/VolumeMapper/test/testCompositeParallelProjection.js
@@ -1,22 +1,21 @@
 import test from 'tape-catch';
 import testUtils from 'vtk.js/Sources/Testing/testUtils';
 
+import 'vtk.js/Sources/Rendering/Misc/RenderingAPIs';
 import vtkColorTransferFunction from 'vtk.js/Sources/Rendering/Core/ColorTransferFunction';
 import vtkHttpDataSetReader from 'vtk.js/Sources/IO/Core/HttpDataSetReader';
-import vtkOpenGLRenderWindow from 'vtk.js/Sources/Rendering/OpenGL/RenderWindow';
 import vtkPiecewiseFunction from 'vtk.js/Sources/Common/DataModel/PiecewiseFunction';
 import vtkRenderWindow from 'vtk.js/Sources/Rendering/Core/RenderWindow';
 import vtkRenderWindowInteractor from 'vtk.js/Sources/Rendering/Core/RenderWindowInteractor';
 import vtkRenderer from 'vtk.js/Sources/Rendering/Core/Renderer';
 import vtkVolume from 'vtk.js/Sources/Rendering/Core/Volume';
 import vtkVolumeMapper from 'vtk.js/Sources/Rendering/Core/VolumeMapper';
-import Constants from 'vtk.js/Sources/Rendering/Core/VolumeMapper/Constants';
 
-import baseline from './testAverageIntensityProjection.png';
+import baseline1 from './testCompositeParallelProjection.png';
 
-test.onlyIfWebGL('Test Average Intensity Projection Volume Rendering', (t) => {
+test('Test Composite Volume Rendering with parallel projection', (t) => {
   const gc = testUtils.createGarbageCollector(t);
-  t.ok('rendering', 'vtkOpenGLVolumeMapper AverageIP');
+  t.ok('rendering', 'vtkOpenGLVolumeMapper CompositeParallelProjection');
   // testUtils.keepDOM();
 
   // Create some control UI
@@ -36,35 +35,27 @@ test.onlyIfWebGL('Test Average Intensity Projection Volume Rendering', (t) => {
 
   const mapper = gc.registerResource(vtkVolumeMapper.newInstance());
   mapper.setSampleDistance(0.7);
-  mapper.setBlendMode(Constants.BlendMode.AVERAGE_INTENSITY_BLEND);
-
   actor.setMapper(mapper);
 
   const reader = vtkHttpDataSetReader.newInstance({ fetchGzip: true });
-
   // create color and opacity transfer functions
   const ctfun = vtkColorTransferFunction.newInstance();
-  ctfun.addRGBPoint(-3024, 0, 0, 0);
-  ctfun.addRGBPoint(-637.62, 1, 1, 1);
-  ctfun.addRGBPoint(700, 1, 1, 1);
-  ctfun.addRGBPoint(3071, 1, 1, 1);
-  ctfun.setMappingRange(500, 3000);
-
+  ctfun.addRGBPoint(0, 85 / 255.0, 0, 0);
+  ctfun.addRGBPoint(95, 1.0, 1.0, 1.0);
+  ctfun.addRGBPoint(225, 0.66, 0.66, 0.5);
+  ctfun.addRGBPoint(255, 0.3, 1.0, 0.5);
   const ofun = vtkPiecewiseFunction.newInstance();
-  ofun.addPoint(-3024, 0);
-  ofun.addPoint(-637.62, 0);
-  ofun.addPoint(700, 0.5);
-  ofun.addPoint(3071, 0.9);
-
+  ofun.addPoint(0.0, 0.0);
+  ofun.addPoint(255.0, 1.0);
   actor.getProperty().setRGBTransferFunction(0, ctfun);
   actor.getProperty().setScalarOpacity(0, ofun);
-  actor.getProperty().setScalarOpacityUnitDistance(0, 4.5);
+  actor.getProperty().setScalarOpacityUnitDistance(0, 3.0);
   actor.getProperty().setInterpolationTypeToFastLinear();
 
   mapper.setInputConnection(reader.getOutputPort());
 
-  // now create something to view it, in this case webgl
-  const glwindow = gc.registerResource(vtkOpenGLRenderWindow.newInstance());
+  // now create something to view it
+  const glwindow = gc.registerResource(renderWindow.newAPISpecificView());
   glwindow.setContainer(renderWindowContainer);
   renderWindow.addView(glwindow);
   glwindow.setSize(400, 400);
@@ -76,16 +67,18 @@ test.onlyIfWebGL('Test Average Intensity Projection Volume Rendering', (t) => {
   interactor.initialize();
   interactor.bindEvents(renderWindowContainer);
 
-  reader.setUrl(`${__BASE_PATH__}/Data/volume/headsq.vti`).then(() => {
+  reader.setUrl(`${__BASE_PATH__}/Data/volume/LIDC2.vti`).then(() => {
     reader.loadData().then(() => {
       renderer.addVolume(actor);
       renderer.resetCamera();
+      renderer.getActiveCamera().zoom(1.5);
+      renderer.getActiveCamera().setParallelProjection(true);
 
       glwindow.captureNextImage().then((image) => {
         testUtils.compareImages(
           image,
-          [baseline],
-          'Rendering/OpenGL/VolumeMapper/testAverageIntensityProjection',
+          [baseline1],
+          'Rendering/Core/VolumeMapper/testCompositeParallelProjection',
           t,
           1.5,
           gc.releaseResources
