+++ Sources/Rendering/Core/VolumeMapper/test/testCompositeParallelProjection.png
