+++ Sources/Rendering/Core/VolumeMapper/test/testComposite_2.png
