+++ Sources/Rendering/Core/VolumeMapper/test/testIntermixedImage.png
