+++ Sources/Rendering/Core/VolumeMapper/test/testIntermixed_2.png
