+++ Sources/Rendering/Core/VolumeMapper/test/testMaximumIntensityProjection.png
