+++ Sources/Rendering/Core/VolumeMapper/test/testMinimumIntensityProjection.js
@@ -1,22 +1,22 @@
 import test from 'tape-catch';
 import testUtils from 'vtk.js/Sources/Testing/testUtils';
 
+import 'vtk.js/Sources/Rendering/Misc/RenderingAPIs';
 import vtkColorTransferFunction from 'vtk.js/Sources/Rendering/Core/ColorTransferFunction';
 import vtkHttpDataSetReader from 'vtk.js/Sources/IO/Core/HttpDataSetReader';
-import vtkOpenGLRenderWindow from 'vtk.js/Sources/Rendering/OpenGL/RenderWindow';
 import vtkPiecewiseFunction from 'vtk.js/Sources/Common/DataModel/PiecewiseFunction';
 import vtkRenderWindow from 'vtk.js/Sources/Rendering/Core/RenderWindow';
 import vtkRenderWindowInteractor from 'vtk.js/Sources/Rendering/Core/RenderWindowInteractor';
 import vtkRenderer from 'vtk.js/Sources/Rendering/Core/Renderer';
 import vtkVolume from 'vtk.js/Sources/Rendering/Core/Volume';
 import vtkVolumeMapper from 'vtk.js/Sources/Rendering/Core/VolumeMapper';
+import Constants from 'vtk.js/Sources/Rendering/Core/VolumeMapper/Constants';
 
-import baseline1 from './testComposite.png';
-import baseline2 from './testComposite_2.png';
+import baseline from './testMinimumIntensityProjection.png';
 
-test.onlyIfWebGL('Test Composite Volume Rendering', (t) => {
+test('Test Minimum Intensity Projection Volume Rendering', (t) => {
   const gc = testUtils.createGarbageCollector(t);
-  t.ok('rendering', 'vtkOpenGLVolumeMapper Composite');
+  t.ok('rendering', 'vtkVolumeMapper MinIP');
   // testUtils.keepDOM();
 
   // Create some control UI
@@ -30,24 +30,31 @@ test.onlyIfWebGL('Test Composite Volume Rendering', (t) => {
   const renderWindow = gc.registerResource(vtkRenderWindow.newInstance());
   const renderer = gc.registerResource(vtkRenderer.newInstance());
   renderWindow.addRenderer(renderer);
-  renderer.setBackground(0.32, 0.34, 0.43);
+  renderer.setBackground(0.3, 0.3, 0.3);
 
   const actor = gc.registerResource(vtkVolume.newInstance());
 
   const mapper = gc.registerResource(vtkVolumeMapper.newInstance());
-  mapper.setSampleDistance(0.7);
+  mapper.setSampleDistance(1.3);
+  mapper.setBlendMode(Constants.BlendMode.MINIMUM_INTENSITY_BLEND);
+
   actor.setMapper(mapper);
 
   const reader = vtkHttpDataSetReader.newInstance({ fetchGzip: true });
+
   // create color and opacity transfer functions
   const ctfun = vtkColorTransferFunction.newInstance();
-  ctfun.addRGBPoint(0, 85 / 255.0, 0, 0);
-  ctfun.addRGBPoint(95, 1.0, 1.0, 1.0);
-  ctfun.addRGBPoint(225, 0.66, 0.66, 0.5);
-  ctfun.addRGBPoint(255, 0.3, 1.0, 0.5);
+  ctfun.addRGBPoint(-3024, 1, 1, 1);
+  ctfun.addRGBPoint(-637.62, 1, 1, 1);
+  ctfun.addRGBPoint(700, 0, 0, 0);
+  ctfun.addRGBPoint(3071, 0, 0, 0);
+
   const ofun = vtkPiecewiseFunction.newInstance();
-  ofun.addPoint(0.0, 0.0);
-  ofun.addPoint(255.0, 1.0);
+  ofun.addPoint(-3024, 0.1);
+  ofun.addPoint(-637.62, 0.1);
+  ofun.addPoint(700, 0.9);
+  ofun.addPoint(3071, 0.9);
+
   actor.getProperty().setRGBTransferFunction(0, ctfun);
   actor.getProperty().setScalarOpacity(0, ofun);
   actor.getProperty().setScalarOpacityUnitDistance(0, 3.0);
@@ -55,8 +62,8 @@ test.onlyIfWebGL('Test Composite Volume Rendering', (t) => {
 
   mapper.setInputConnection(reader.getOutputPort());
 
-  // now create something to view it, in this case webgl
-  const glwindow = gc.registerResource(vtkOpenGLRenderWindow.newInstance());
+  // now create something to view it
+  const glwindow = gc.registerResource(renderWindow.newAPISpecificView());
   glwindow.setContainer(renderWindowContainer);
   renderWindow.addView(glwindow);
   glwindow.setSize(400, 400);
@@ -68,19 +75,17 @@ test.onlyIfWebGL('Test Composite Volume Rendering', (t) => {
   interactor.initialize();
   interactor.bindEvents(renderWindowContainer);
 
-  reader.setUrl(`${__BASE_PATH__}/Data/volume/LIDC2.vti`).then(() => {
+  reader.setUrl(`${__BASE_PATH__}/Data/volume/headsq.vti`).then(() => {
     reader.loadData().then(() => {
       renderer.addVolume(actor);
       renderer.resetCamera();
-      renderer.getActiveCamera().zoom(1.5);
-      renderer.getActiveCamera().elevation(70);
-      renderer.resetCameraClippingRange();
+      renderer.getActiveCamera().elevation(-120);
 
       glwindow.captureNextImage().then((image) => {
         testUtils.compareImages(
           image,
-          [baseline1, baseline2],
-          'Rendering/OpenGL/VolumeMapper/testComposite',
+          [baseline],
+          'Rendering/Core/VolumeMapper/testMinimumIntensityProjection',
           t,
           1.5,
           gc.releaseResources
