+++ Sources/Rendering/Core/VolumeMapper/test/testMinimumIntensityProjection.png
