--- Sources/Rendering/OpenGL/VolumeMapper/test/testCompositeParallelProjection.js
@@ -1,93 +1,0 @@
-import test from 'tape-catch';
-import testUtils from 'vtk.js/Sources/Testing/testUtils';
-
-import vtkColorTransferFunction from 'vtk.js/Sources/Rendering/Core/ColorTransferFunction';
-import vtkHttpDataSetReader from 'vtk.js/Sources/IO/Core/HttpDataSetReader';
-import vtkOpenGLRenderWindow from 'vtk.js/Sources/Rendering/OpenGL/RenderWindow';
-import vtkPiecewiseFunction from 'vtk.js/Sources/Common/DataModel/PiecewiseFunction';
-import vtkRenderWindow from 'vtk.js/Sources/Rendering/Core/RenderWindow';
-import vtkRenderWindowInteractor from 'vtk.js/Sources/Rendering/Core/RenderWindowInteractor';
-import vtkRenderer from 'vtk.js/Sources/Rendering/Core/Renderer';
-import vtkVolume from 'vtk.js/Sources/Rendering/Core/Volume';
-import vtkVolumeMapper from 'vtk.js/Sources/Rendering/Core/VolumeMapper';
-
-import baseline1 from './testCompositeParallelProjection.png';
-
-test.onlyIfWebGL(
-  'Test Composite Volume Rendering with parallel projection',
-  (t) => {
-    const gc = testUtils.createGarbageCollector(t);
-    t.ok('rendering', 'vtkOpenGLVolumeMapper CompositeParallelProjection');
-    // testUtils.keepDOM();
-
-    // Create some control UI
-    const container = document.querySelector('body');
-    const renderWindowContainer = gc.registerDOMElement(
-      document.createElement('div')
-    );
-    container.appendChild(renderWindowContainer);
-
-    // create what we will view
-    const renderWindow = gc.registerResource(vtkRenderWindow.newInstance());
-    const renderer = gc.registerResource(vtkRenderer.newInstance());
-    renderWindow.addRenderer(renderer);
-    renderer.setBackground(0.32, 0.34, 0.43);
-
-    const actor = gc.registerResource(vtkVolume.newInstance());
-
-    const mapper = gc.registerResource(vtkVolumeMapper.newInstance());
-    mapper.setSampleDistance(0.7);
-    actor.setMapper(mapper);
-
-    const reader = vtkHttpDataSetReader.newInstance({ fetchGzip: true });
-    // create color and opacity transfer functions
-    const ctfun = vtkColorTransferFunction.newInstance();
-    ctfun.addRGBPoint(0, 85 / 255.0, 0, 0);
-    ctfun.addRGBPoint(95, 1.0, 1.0, 1.0);
-    ctfun.addRGBPoint(225, 0.66, 0.66, 0.5);
-    ctfun.addRGBPoint(255, 0.3, 1.0, 0.5);
-    const ofun = vtkPiecewiseFunction.newInstance();
-    ofun.addPoint(0.0, 0.0);
-    ofun.addPoint(255.0, 1.0);
-    actor.getProperty().setRGBTransferFunction(0, ctfun);
-    actor.getProperty().setScalarOpacity(0, ofun);
-    actor.getProperty().setScalarOpacityUnitDistance(0, 3.0);
-    actor.getProperty().setInterpolationTypeToFastLinear();
-
-    mapper.setInputConnection(reader.getOutputPort());
-
-    // now create something to view it, in this case webgl
-    const glwindow = gc.registerResource(vtkOpenGLRenderWindow.newInstance());
-    glwindow.setContainer(renderWindowContainer);
-    renderWindow.addView(glwindow);
-    glwindow.setSize(400, 400);
-
-    // Interactor
-    const interactor = vtkRenderWindowInteractor.newInstance();
-    interactor.setStillUpdateRate(0.01);
-    interactor.setView(glwindow);
-    interactor.initialize();
-    interactor.bindEvents(renderWindowContainer);
-
-    reader.setUrl(`${__BASE_PATH__}/Data/volume/LIDC2.vti`).then(() => {
-      reader.loadData().then(() => {
-        renderer.addVolume(actor);
-        renderer.resetCamera();
-        renderer.getActiveCamera().zoom(1.5);
-        renderer.getActiveCamera().setParallelProjection(true);
-
-        glwindow.captureNextImage().then((image) => {
-          testUtils.compareImages(
-            image,
-            [baseline1],
-            'Rendering/OpenGL/VolumeMapper/testCompositeParallelProjection',
-            t,
-            1.5,
-            gc.releaseResources
-          );
-        });
-        renderWindow.render();
-      });
-    });
-  }
-);
