+++ Sources/tests.js
@@ -60,6 +60,7 @@ import './Rendering/Core/ImageMapper/test/testImage';
 import './Rendering/Core/ImageMapper/test/testImageNearestNeighbor';
 import './Rendering/Core/ImageMapper/test/testImageColorTransferFunction';
 import './Rendering/Core/Mapper/test/testEdgeVisibility';
+import './Rendering/Core/Mapper/test/testInterpolateScalarsBeforeMapping';
 import './Rendering/Core/Mapper/test/testVectorComponent';
 import './Rendering/Core/PointPicker/test/testPointPicker';
 import './Rendering/Core/Prop3D/test/testUserMatrix';
@@ -67,25 +68,24 @@ import './Rendering/Core/RenderWindow/test/testMultipleRenderers';
 import './Rendering/Core/SphereMapper/test/testDisableScalarColoring';
 import './Rendering/Core/SphereMapper/test/testSphere';
 import './Rendering/Core/StickMapper/test/testStick';
+import './Rendering/Core/VolumeMapper/test/testAverageIntensityProjection';
+import './Rendering/Core/VolumeMapper/test/testComposite';
+import './Rendering/Core/VolumeMapper/test/testCompositeParallelProjection';
+import './Rendering/Core/VolumeMapper/test/testComposite16Bit';
+import './Rendering/Core/VolumeMapper/test/testIntermixed';
+import './Rendering/Core/VolumeMapper/test/testIntermixedImage';
+import './Rendering/Core/VolumeMapper/test/testMaximumIntensityProjection';
+import './Rendering/Core/VolumeMapper/test/testMinimumIntensityProjection';
 import './Rendering/Misc/GenericRenderWindow/test/testGenericRenderWindowCreateDelete';
 import './Rendering/OpenGL/ImageMapper/test/testImageCroppingPlanes';
 import './Rendering/OpenGL/PolyDataMapper/test/testAddShaderReplacements';
 import './Rendering/OpenGL/PolyDataMapper/test/testClearShaderReplacements';
 import './Rendering/OpenGL/PolyDataMapper/test/testClippingPlanes';
-import './Rendering/OpenGL/PolyDataMapper/test/testInterpolateScalarsBeforeMapping';
 import './Rendering/OpenGL/Skybox/test/testSkybox';
 import './Rendering/OpenGL/Skybox/test/testSkyboxBackground';
 import './Rendering/OpenGL/Texture/test/testCreateCubeFromRawTexture';
-import './Rendering/OpenGL/VolumeMapper/test/testComposite';
-import './Rendering/OpenGL/VolumeMapper/test/testCompositeParallelProjection';
-import './Rendering/OpenGL/VolumeMapper/test/testComposite16Bit';
-import './Rendering/OpenGL/VolumeMapper/test/testIntermixed';
-import './Rendering/OpenGL/VolumeMapper/test/testIntermixedImage';
 import './Rendering/OpenGL/VolumeMapper/test/testLighting';
-import './Rendering/OpenGL/VolumeMapper/test/testMaximumIntensityProjection';
-import './Rendering/OpenGL/VolumeMapper/test/testMinimumIntensityProjection';
 import './Rendering/OpenGL/VolumeMapper/test/testProportionalComponents';
-import './Rendering/OpenGL/VolumeMapper/test/testAverageIntensityProjection';
 import './Testing/testAlgorithm';
 import './Testing/testIf';
 import './Testing/testMacro';
